# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailsResponse"
                }
              }
            }
          }
        },
        "summary": "Create a new task",
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "description": "Returns tasks based on user role. Admin sees all tasks, Manager sees their created tasks, Member sees assigned tasks.",
        "operationId": "TasksController_getTasks",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "DONE"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
              ]
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTaskResponse"
                }
              }
            }
          }
        },
        "summary": "Get all tasks",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "description": "Returns task details. Fields available depend on user role: Admin (all fields), Manager (no creator), Member (no assignee).",
        "operationId": "TasksController_getTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailsResponse"
                }
              }
            }
          }
        },
        "summary": "Get a task by id",
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailsResponse"
                }
              }
            }
          }
        },
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/status": {
      "patch": {
        "operationId": "TasksController_updateTaskStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/projects": {
      "get": {
        "description": "Returns projects based on user role. Admin sees all projects, Manager sees their created projects, Member sees projects they are participated.",
        "operationId": "ProjectsController_getAllProjects",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectResponse"
                }
              }
            }
          }
        },
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "description": "Returns task details. Fields available depend on user role: Admin (all fields), Manager (no creator), Member (no assignee).",
        "operationId": "ProjectsController_getProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailsResponse"
                }
              }
            }
          }
        },
        "summary": "Get a project by id",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_deleteProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tasks Management API",
    "description": "API for Tasks Management",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Task title"
          },
          "description": {
            "type": "string",
            "description": "Task description"
          },
          "projectId": {
            "type": "string",
            "description": "Project of task"
          },
          "assigneeId": {
            "type": "string",
            "description": "Task assignee"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "Task due date"
          }
        },
        "required": [
          "title",
          "projectId",
          "assigneeId"
        ]
      },
      "ProjectInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "createdAt",
          "ownerId"
        ]
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "TaskDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "projectId": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          },
          "creator": {
            "description": "Available for Admin and Member roles",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfo"
              }
            ]
          },
          "assignee": {
            "description": "Available for Admin and Manager roles",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfo"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "projectId",
          "creatorId",
          "assigneeId",
          "createdAt",
          "updatedAt",
          "project"
        ]
      },
      "TaskListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "projectId": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          },
          "creator": {
            "$ref": "#/components/schemas/UserInfo"
          },
          "assignee": {
            "$ref": "#/components/schemas/UserInfo"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "projectId",
          "creatorId",
          "assigneeId",
          "createdAt",
          "updatedAt",
          "project"
        ]
      },
      "PaginatedTaskResponse": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskListItem"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "tasks",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Task title"
          },
          "description": {
            "type": "string",
            "description": "Task description"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "Task due date"
          }
        }
      },
      "UpdateTaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "DONE"
            ]
          }
        }
      },
...
      "OwnerInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "CountsInProjectItem": {
        "type": "object",
        "properties": {
          "members": {
            "type": "number",
            "description": "Availabe for all users"
          },
          "tasks": {
            "type": "number",
            "description": "Availabe for all users"
          }
        }
      },
      "TaskInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "createdAt"
        ]
      },
      "ProjectListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerInfo"
          },
          "_count": {
            "description": "Numbers of members and tasks",
            "allOf": [
              {
                "$ref": "#/components/schemas/CountsInProjectItem"
              }
            ]
          },
          "task": {
            "$ref": "#/components/schemas/TaskInfo"
          }
        },
        "required": [
          "id",
          "name",
          "ownerId",
          "createdAt",
          "_count"
        ]
      },
      "PaginatedProjectResponse": {
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectListItem"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "projects",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
      "MemberInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "userId",
          "user"
        ]
      },
      "ProjectDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "description": "Availabe for Admin and Member",
            "allOf": [
              {
                "$ref": "#/components/schemas/OwnerInfo"
              }
            ]
          },
          "members": {
            "description": "Available for Admin and Manager",
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberInfo"
              }
            ]
          },
          "tasks": {
            "description": "Avaialbe for all users",
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskInfo"
              }
            ]
          },
          "_count": {
            "type": "object",
            "description": "Avaialbe for members"
          }
        },
        "required": [
          "id",
          "name",
          "ownerId",
          "createdAt"
        ]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}